fix(entry): resolve 6 code review findings for story 1-5
5-layer adversarial review (Blind Hunter, Edge Case Hunter, Acceptance
Auditor, Architecture Guardian, iOS/SwiftUI Guardian) produced 36 raw
findings → 11 unique after dedup → 6 patches, 5 deferred, 14 dismissed.

CRITICAL fixes:
- GeometryReader/LazyVGrid circular layout: interpose VStack(spacing: 0)
- displayAmount Double → Decimal: enforce "no floating-point for money"

WARNING fixes:
- Add cap boundary test at 999_999 → 9_999_999
- maxBeforeAppend: private static let + clarified comment
- @MainActor at test class level (not per-method)
- NumpadView preview: add .frame(height: 300)

All 50 tests pass. Story 1-5 status → done.

+++ CashOut/Utilities/Extensions/Int64+Currency.swift
@@ -4,7 +4,7 @@ extension Int64 {
     /// Formats satang as THB string (e.g., 1250 → "฿12.50").
     /// Hardcoded to th_TH / THB — intentional for this personal-use, THB-only app.
     var displayAmount: String {
-        let baht = Double(self) / 100.0
+        let baht = Decimal(self) / 100
         return baht.formatted(.currency(code: "THB").locale(Locale(identifier: "th_TH")))
     }
 }
